// src/generated/query/zod/metroMirrorRelationshipsResponse.gen.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * ABCo API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';

export const MetroMirrorRelationshipsResponse = zod.object({
  "provider_id": zod.string(),
  "volumes": zod.array(zod.object({
  "name": zod.string(),
  "status": zod.enum(['ok', 'not_mirrored', 'ambiguous']),
  "auxiliary_name": zod.union([zod.string(),zod.null()]).exactOptional(),
  "consistency_group_id": zod.union([zod.string(),zod.null()]).exactOptional()
})),
  "consistency_group_id": zod.union([zod.string(),zod.null()]).exactOptional(),
  "warning": zod.union([zod.string(),zod.null()]).exactOptional()
})

export type MetroMirrorRelationshipsResponse = zod.input<typeof MetroMirrorRelationshipsResponse>;
export type MetroMirrorRelationshipsResponseOutput = zod.output<typeof MetroMirrorRelationshipsResponse>;